#526 Merge this feature to current command datamodel.
