+++ cmd/sysl/cmd_datamodel.go
@@ -47,6 +47,44 @@ func GenerateDataModels(datagenParams *CmdContextParamDatagen,
 	return outmap, nil
 }
 
+/**
+ * It is added to help reviewing generated data model with sysl
+ * file produced by command import. Generate data model diagrams using the following command:
+ * sysl reviewdata --root=/Users/guest/data -t Test -o Test.png Test
+ * sysl reviewdatamodel --root=/Users/guest/data -t Test -o Test.png Test.sysl
+ */
+func GenerateDataModelsView(datagenParams *CmdContextParamDatagen,
+	model *sysl.Module, logger *logrus.Logger) (map[string]string, error) {
+	outmap := make(map[string]string)
+
+	logger.Debugf("title: %s\n", datagenParams.title)
+	logger.Debugf("output: %s\n", datagenParams.output)
+
+	spclass := constructFormatParser("", datagenParams.classFormat)
+
+	apps := model.GetApps()
+	for appName := range apps {
+		app := apps[appName]
+		outputDir := datagenParams.output
+		if strings.Contains(outputDir, "%(appname)") {
+			of := MakeFormatParser(datagenParams.output)
+			outputDir = of.FmtOutput(appName, "", app.GetLongName(), app.GetAttrs())
+		}
+		var stringBuilder strings.Builder
+		if app != nil {
+			dataParam := &DataModelParam{
+				mod:   model,
+				app:   app,
+				title: datagenParams.title,
+			}
+			v := MakeDataModelView(spclass, dataParam.mod, &stringBuilder, dataParam.title, "")
+			outmap[outputDir] = v.GenerateDataView(dataParam)
+		}
+	}
+
+	return outmap, nil
+}
+
 func generateDataModel(pclass ClassLabeler, outmap map[string]string, mod *sysl.Module,
 	stmts []*sysl.Statement, title, project, outDir string) {
 	apps := mod.GetApps()
@@ -100,9 +138,38 @@ func (p *datamodelCmd) Configure(app *kingpin.Application) *kingpin.CmdClause {
 }
 
 func (p *datamodelCmd) Execute(args ExecuteArgs) error {
-	outmap, err := GenerateDataModels(&p.CmdContextParamDatagen, args.Modules[0], args.Logger)
+	var model = args.Modules[0]
+	var outmap map[string]string
+	var err error
+	if !projectManner(model) {
+		// The sysl file is not project manner
+		outmap, err = GenerateDataModelsView(&p.CmdContextParamDatagen, model, args.Logger)
+	} else {
+		// Sysl file is project manner
+		outmap, err = GenerateDataModels(&p.CmdContextParamDatagen, model, args.Logger)
+	}
+
 	if err != nil {
 		return err
 	}
 	return p.GenerateFromMap(outmap, args.Filesystem)
+}
+
+/**
+ * Check if sysl is project manner or not
+ */
+func projectManner(model *sysl.Module) bool {
+	var onlyApp *sysl.Application
+	apps := model.GetApps()
+	if len(apps) == 1 {
+		for _, app := range apps {
+			onlyApp = app
+		}
+	}
+
+	if onlyApp != nil && onlyApp.GetEndpoints() == nil {
+		return false
+	}
+
+	return true
 }
